+++ packages/scripts/src/projects/background.ts
@@ -16,7 +16,7 @@ import {
 import semver_gt from 'semver/functions/gt';
 import semver_valid from 'semver/functions/valid';
 import { CommonProject } from './common';
-import { CXProject, definedProjects, ICourseProject, IcveMoocProject, ZHSProject, ZJYProject } from '..';
+import { CXProject, definedProjects, ICourseProject, IcveMoocProject, YKTProject, ZHSProject, ZJYProject } from '..';
 import { RenderScript } from '../render';
 import { SearchInfosElement } from '../elements/search.infos';
 import { $render } from '../utils/render';
@@ -943,7 +943,9 @@ export const BackgroundProject = Project.create({
 					[ZJYProject.scripts.dispatcher, ZJYProject.scripts.study],
 					ZJYProject.scripts.work,
 					IcveMoocProject.scripts.study,
-					IcveMoocProject.scripts.work
+					IcveMoocProject.scripts.work,
+					YKTProject.scripts.ai,
+					YKTProject.scripts.v2_study
 				]
 					.map((m) => {
 						const url = window.location.href;