+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## [4.13.7](https://github.com/ocsjs/ocsjs/compare/4.13.6...4.13.7) (2026-05-14)
+
+
+### Features
+
+* **script:** 新增雨课堂V2学习内容自动看视频、和讨论功能 ([d8095b6](https://github.com/ocsjs/ocsjs/commit/d8095b63104706cb2ac46d899e573da6c778eafa))
+
+
+
+## [4.13.6](https://github.com/ocsjs/ocsjs/compare/4.13.4...4.13.6) (2026-05-12)
+
+
+### Bug Fixes
+
+* **script:** 适配智慧树-智慧课程-AI学伴掌握度无法答题的问题 ([ce6262f](https://github.com/ocsjs/ocsjs/commit/ce6262fabaa3edcf38056863c7c30319ff5ece48))
+
+
+### Performance Improvements
+
+* **script:** 修改题库配置按钮文案，优化题库配置界面文案，主次分明 ([3d237ea](https://github.com/ocsjs/ocsjs/commit/3d237ea7c2284c7560ef6fe8801b5491ed3e6dfc))
+
+
+
+## [4.13.4](https://github.com/ocsjs/ocsjs/compare/4.13.2...4.13.4) (2026-04-23)
+
+
+### Features
+
+* **script:** 兼容超星长时阅读任务点（限制翻页时间） ([ac34ba2](https://github.com/ocsjs/ocsjs/commit/ac34ba2d543afe9dc09c1803b558f7b37dbce095))
+* **script:** 添加超星域名支持：jnzyjsxy.cn ([a27cb49](https://github.com/ocsjs/ocsjs/commit/a27cb49b9111e3317120c33eff75981de4b19ed8))
+
+
+
+## [4.13.2](https://github.com/ocsjs/ocsjs/compare/4.12.39...4.13.2) (2026-04-22)
+
+
+### Bug Fixes
+
+* **script:** 修复智慧树26年上半年studywisdomh5复习模式错误的问题 ([2e82a0a](https://github.com/ocsjs/ocsjs/commit/2e82a0a8ea4cca4265ba592336a2762bf156ec08))
+
+
+### Features
+
+* **script:** 兼容雨课堂AI学伴自动学习 ([c96f4db](https://github.com/ocsjs/ocsjs/commit/c96f4db7d949df6f3fc983a6ae4783ece9b19783))
+
+
+
 ## [4.12.39](https://github.com/ocsjs/ocsjs/compare/4.12.38...4.12.39) (2026-04-20)